renamed petclinic to kidclinic
