+++ src/main/java/org/springframework/samples/kidclinic/model/NamedEntity.java
@@ -13,43 +13,36 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.samples.petclinic.model;
+package org.springframework.samples.kidclinic.model;
 
 import javax.persistence.Column;
 import javax.persistence.MappedSuperclass;
 
-import org.hibernate.validator.constraints.NotEmpty;
 
 /**
- * Simple JavaBean domain object representing an person.
+ * Simple JavaBean domain object adds a name property to <code>BaseEntity</code>. Used as a base class for objects
+ * needing these properties.
  *
  * @author Ken Krebs
+ * @author Juergen Hoeller
  */
 @MappedSuperclass
-public class Person extends BaseEntity {
+public class NamedEntity extends BaseEntity {
 
-    @Column(name = "first_name")
-    @NotEmpty
-    private String firstName;
+    @Column(name = "name")
+    private String name;
 
-    @Column(name = "last_name")
-    @NotEmpty
-    private String lastName;
-
-    public String getFirstName() {
-        return this.firstName;
+    public String getName() {
+        return this.name;
     }
 
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
+    public void setName(String name) {
+        this.name = name;
     }
 
-    public String getLastName() {
-        return this.lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
+    @Override
+    public String toString() {
+        return this.getName();
     }
 
 }
