+++ src/main/java/org/springframework/samples/kidclinic/owner/OwnerController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.samples.petclinic.owner;
+package org.springframework.samples.kidclinic.owner;
 
 import java.util.Collection;
 import java.util.Map;
