+++ src/main/java/org/springframework/samples/kidclinic/owner/PetType.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.samples.petclinic.owner;
+package org.springframework.samples.kidclinic.owner;
 
 import javax.persistence.Entity;
 import javax.persistence.Table;
 
-import org.springframework.samples.petclinic.model.NamedEntity;
+import org.springframework.samples.kidclinic.model.NamedEntity;
 
 /**
  * @author Juergen Hoeller
