+++ src/main/java/org/springframework/samples/kidclinic/vet/Vets.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.samples.petclinic.vet;
+package org.springframework.samples.kidclinic.vet;
 
 import java.util.ArrayList;
 import java.util.List;
