+++ src/main/java/org/springframework/samples/kidclinic/visit/Visit.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.samples.petclinic.visit;
+package org.springframework.samples.kidclinic.visit;
 
 import java.util.Date;
 
@@ -25,7 +25,7 @@ import javax.persistence.TemporalType;
 
 import org.hibernate.validator.constraints.NotEmpty;
 import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.samples.petclinic.model.BaseEntity;
+import org.springframework.samples.kidclinic.model.BaseEntity;
 
 /**
  * Simple JavaBean domain object representing a visit.
